Trivial: Make vertical white space after license header consistent
Vertical white space between license header and the actual code is not consistent
across files. It looks like majority of the files leaves a single blank line
after license header. So make it consistent except for those exceptional cases
where the actual code starts with a function or class definition.

Change-Id: If6679ab19dc2d7a9df93e076db3019abc1e5b86b