Add deprecated name test case
This is being separated from the change to deprecate the DEFAULT
grouped options because it is exposing a python 3 incompatibility
in oslotest and I don't want to block the release on something this
minor.  Once oslotest is fixed we can merge this to make sure
the deprecated options continue working.

Change-Id: Iabf5ad6ad1e7ad9a426783592f23f58452484b32