Merge fileutils from oslo-incubator
We need to get rid of our copy of log.py in oslo.concurrency.

oslo.concurrency code needs fileutils from oslo-incubator.
fileutils needs the common copy of log.py. What happens when we
import oslo.concurrency into nova, we end up with duplicate
options for default_log_levels. So to unravel the knots,
we switch fileutils over to python logging and hence remove
the need for common copy of log.py in oslo.concurrency.

Fixed the import for excutils as well since that change
has already been made in the oslo-incubator copy of fileutils

Closes-Bug: #1385492
Change-Id: I5310bf8a6584ac4f71b607d30b9dae0b12441d41