lockutils-wrapper cleanup
There were a couple of issues with the lockutils wrapper.  Most
importantly, lockutils silently did nothing if it was called in the
same way as the old incubator lockutils module.  This had the
potential to cause problems when simply doing a find and replace to
migrate to oslo.concurrency.  The problem is fixed by making
lockutils raise an exception if it is called directly.

There was also a name mismatch in the docstring for the console
entry point, which is also fixed in this change.

Closes-Bug: 1386734
Change-Id: I8868820ca314eb8d6fee83cc66fea886c1e74e27