test compatibility for old imports
The following test is for compatibility for the namespace move. Having
this test in the oslo.concurrency tree will ensure that not only do
the imports functionally work, but that hacking can pass on this test,
which ensures that hacking is compatible after the namespace change.

The previous patches to oslo.concurrency extensively used # noqa,
which masked out the issue for oslo.concurrency, thus allowing the
release to break other parts of OpenStack.

Change-Id: Icd118a00a8ec0d9fb0bd9d60826676eec2a08f47
Related-Bug: #1398472