Port processutils to Python 3
On Python 3, execute() and ssh_execute() now return stdout and stderr as
Unicode. os.fsdecode() is used to stdout and stderr, it uses the locale
encoding with the surrogateescape error handler.
subprocess.Popen(universal_encoding=True) uses the same encoding and
error handler.

This change breaks the backward compatibility, but I believe that
Unicode is more convinient in Python 3 and it is what users expect. In
practice, there is no known (OpenStack) application compatible with
Python 3 using oslo_concurrency.processutils yet.

Add new tests.

This changeset does not change the behaviour on Python 2, only Python 3
is impacted.

Change-Id: Ibb510642c1440e2fbe7ad53d797bc7c32a3cbddf