Add eventlet test check to new tests __init__.py
This is used to allow us to test both with and without eventlet.
It wasn't copied to the new tests directory in the namespace move,
which means that once the old tests go away eventlet won't be
monkey patching even when we try to make it.  Also, there's a race
depending on whether the top-level tests are run first, or the new
package ones.

Change-Id: Ie2705c3d21625e594471f27eefc3ac3147390cb7