Add binary parameter to execute and ssh_execute
Add an optional binary parameter to execute() and ssh_execute()
functions of oslo_concurrency.processutils. If binary is True, stdout
and stderr are returned as byte strings on Python 2 and Python 3.

When Nova will be ported to Python 3, binary=True will be needed to get
binary outputs (like an encryption key, see the bug #1410348).

Add new tests.

This changeset does not change the default behaviour.

Change-Id: I283cadf04781942b65f211237070cc2354939cdb
Related-Bug: 1410348