Ignore prlimit argument on Windows
The built in 'resource' module is used in order to apply process
resource limits. This is not available on Windows.

For the moment, we'll ignore this argument and log a warning
message. In the future, we may provide a portable implementation
for this feature.

Change-Id: I6a82916eb68fbf8737b45a65f71cae1f835d12d1
Closes-Bug: #1627766
(cherry picked from commit 70ff5511053bc7070c071f46978b1ff59b89c4b5)