Automatically convert process_input to bytes
On py35, if process_input is a string we end with an Error:
TypeError: memoryview: a bytes-like object is required, not 'str'

A whole lot of code in os-brick/cinder/neutron currently need
to be fixed. Hardest problem is tracking down every instance of
this problem and trying to test/fix it at source. See example
of cinder/os-brick/oslo.rootwrap/oslo.privep code problem here:
http://logs.openstack.org/43/418643/2/check/gate-rally-dsvm-py35-cinder-nv/6efa7b5/logs/screen-n-cpu.txt.gz?level=ERROR

It's just better to fix it in one spot and pass the correct
thing to subprocess.communicate

Change-Id: I88d1510a7ba4c020f73452f0b80e996c22b1edf1