Don't warn on lock removal if file doesn't exist
There may be cases were library users want to request the removal of an
external lock that may not have been created and don't want to pollute
the logs when the file is not present.

We raise the log level from info to warning, but we don't log it if the
file didn't exist, since the end result the caller wanted is there.

Change-Id: I5ce8be34c9f2c4c59ea99dabc6760c3300f743a3