Stop configuring install_command in tox.
Currently, we are overriding 'install_command' to use 'pip'. This is
considered poor behavior and 'python -m pip' should be used instead:

https://snarky.ca/why-you-should-use-python-m-pip/

It turns out that this is the the default value provided by tox:

https://tox.readthedocs.io/en/latest/config.html#conf-install_command

So we can remove the line and simply use the default value.

Also these changes introduce a new explicit env var related to eventlet
testing to avoid misunderstand.

Change-Id: I03b9b179c4cb6ea5a9a8d8c74af3502a0243ace6