[stable-only] Cap virtualenv for py37
py37 test started to fail with latest virtualenv (20.16.4) as it
bundles setuptools 65.3.0, which causes installation problems under
py37. (Note: with py38 the problem does not occur).

This patch sets virtualenv<20.16.4 to unblock the gate as
virtualenv 20.16.3 bundles setuptools 63.4.1, with which py37 gate
works well.

NOTE(stephenfin): Modified to remove the failing l-c job.

Change-Id: I8c457d06b79cacc770db62eac1d8e4459052d5ba