Run mypy from tox
This avoids the need to duplicate our dependency list in multiple
places and allows us to take advantage of tox's dependency management
infrastructure, to ensure we always get the latest and greatest version
of a package allowed by upper-constraints.

This change highlights a few casts that are no longer needed since
oslo.i18n is now typed.

Change-Id: Idf482dbc252bd898ba74a33972d035584b302772
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>